requirements/base.txt
asgiref==3.5.2
Django==4.1
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-environ==0.9.0
djangorestframework==3.13.1
django-phonenumber-field==7.0.0
phonenumbers==8.12.55
drf-yasg==1.20.0
Pillow==9.0.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2